requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
pillow==10.4.0
python-multipart==0.0.9
